Update patch set 1
Patch Set 1:

I had to go through the security review and the lengthy discussion with you all to make it in, but the lack of remote module support made me stop working on the project. I could not build the prototype anymore without the remote module system.

The project was blocked by two things -
1. Unclear future of remote module support.
2. ES6 module conversion in DevTools.

What's the current situation? The core of this project is completed by ~90% and the only remaining step is to integrate the app to the new ES6 system. (which I find quite difficult because it was a moving target several months ago)

Patch-set: 1
Reviewer: Gerrit User 1130280 <1130280@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>