Update patch set 2
Patch Set 2:

It's nice, I like that it adds a bit more contrast to the input everywhere and it looks more interactive.

On another note: don't you think that the original bug is rather vice versa? The horizontal top bar in the coverage panel is the only bar in DevTools that has a white background. Probably a visual bug? That results in the input being invisible. Shouldn't that top bar be grey?

In general I don't like grey interfaces and prefer white (also more Googley), so this is a great fix!

Patch-set: 2
Reviewer: Gerrit User 1419292 <1419292@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>