Let NetworkManager.NetworkDispatcher implements the correct interface
This ports the NetworkDispatcher over from the old parameter arguments
to the object notation. It updates the InspectorBackend to call the
dispatcher with an object, if that dispatcher accepts that.

We use TypeScript to enforce that any dispatcher that implements
the ProtocolProxyApi, correctly returns `true` and thus the
InspectorBackend will not call an old dispatcher with the object
notation.

R=jacktfranklin@chromium.org,aerotwist@chromium.org
CC=sigurds@chromium.org,janscheffler@chromium.org

Fixed: 1081686
Change-Id: Ia315039a9aa8e29ec8553786db1b52dde8089dc6