Add simple touch scroll support in the request list
Unlike other scrollable areas in DevTools, the request list in the
network panel does not allow scrolling with touch.
The reason is that this list requires JS to scroll properly, and the
code that manages this only handles mouse events.

This change adds simple touch event handling that mostly reuses
existing code. It doesn't support  touch-based scrolling exactly
like other areas of devtools unfortunately, like momentum. I'm not
sure it's worth the effort to do this.

Regarding testing, I think a new end to end test would be best, and
I started writing one, but the lack of touch event support in puppeteer
didn't make this super easy. So this change only contains production
code for now.

Bug: 1083203
Change-Id: Ia80cd284847773cd96764f1daf4bc9e7ec1724e1