Include workaround for TypeScript compiler bug
The TypeScript compiler has a bug where, when you implement an
interface in JSDoc which is defined in a `.d.ts` file which
uses a namespace, it generates an invalid declaration file [1].

To workaround this bug, we have to add global interfaces that
are essentially aliases of the actual interfaces. As a result,
the TypeScript compiler no longer fails with resolving.

Update Closure accordingly such that we can use the correct
type to work in both Closure and TypeScript.

[1]: https://github.com/microsoft/TypeScript/issues/38640

R=jacktfranklin@chromium.org,aerotwist@chromium.org
CC=sigurds@chromium.org,janscheffler@chromium.org

Bug: 1081686, 1011811
Change-Id: I088a4520c90ce0873cf9850a8ed65f9c8dd896ef