Add verify_lib_check to ts_library
Previously, we were gating the `skipLibCheck` on `is_debug=true`.
However, this introduced the problem where CQ runs would fail on
compilation, but local runs succeeding. This confused developers
and required handling a different GN output directory with the GN
flag `is_debug=false`.

But, we don't want to lose the performance benefit related to not
unnecessarily checking all `.d.ts` files all the time. So we have
to find a middleground.

The solution is to add 1 `ts_library` target in `front_end` which
essentially verifies that the `.d.ts` are all in good order. This
means we only do the check once, irrespective the number of
`ts_library` targets in the frontend.

DISABLE_THIRD_PARTY_CHECK=TypeScript fix
R=jacktfranklin@chromium.org
CC=sigurds@chromium.org

Bug: 1011811
Change-Id: Ide2d3a5935fdc34dc79572d65c9e5ce92d78a9a8