[Dom Storage]: Handle undefined storage case
Issue:
- DomStorageModel.clearForOrigin does not protect against the case where _storages[key] is undefined

Changes:
- Adding !storage check to clearForOrigin

Description:
There are a ton of moving parts in this code.  Navigation, frame addition and frame removal all trigger a series of separate events that change the main security origin, remove unused origins and add new origins.
The failure is occurs when ClearStorageView._securityOrigin has a mismatch of what origins are in the DOMStorageModel._storages map at the time the user clears the site data.

I believe that adding a simple storage check in clearForOrigin

Bug: 1084631
Change-Id: I7d56a62b1b3ea3e5499f495b1820e7cb789ed0f0