[Dom Storage]: Handle undefined storage case
Issue:
- DomStorageModel.clearForOrigin does not protect against the case where _storages[key] is undefined

Changes:
- Adding !storage check to clearForOrigin

Description:
There are a ton of moving parts in this code.  Navigation, frame addition and frame removal all trigger a series of separate events that change the main security origin, remove unused origins and add new origins.
The failure is occurs when ClearStorageView._securityOrigin has a mismatch of what origins are in the DOMStorageModel._storages map at the time the user clears the site data.
ClearStorageView._updateOrigin (which sets _securityOrigin) is called many times per navigation, so it easy to see that it could be out of sync when a user clears the storage

I believe that simply adding an undefined check in clearForOrigin is the best solution because...
- It is a simple, targeted fix.
    - Adding flags throughout multiple files to ensure that the origins are synced at the time a user tries to clear the storage will effectively do the same thing as an undefined check.
- This will not cause unexpected behavior
    - We don't need to clear the storage of an origin that isn't stored in the _storages map

Bug: 1084631
Change-Id: I7d56a62b1b3ea3e5499f495b1820e7cb789ed0f0