Update patch set 2
Patch Set 2:

I have discussed the issue with Peter Müller and it turns out there is no standard approach for cursor:pointer in DevTools. We agreed that it is probably best to revert the change for now and Peter
will look into various inconsistencies in the user of cursor:pointer that we found in Chrome UI and DevTools UI and figure out the common approach.

Patch-set: 2
Reviewer: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>