Update patch set 7
Patch Set 7:

(1 comment)

I think these are both good points to consider.

Now that I think about it, I think there are pros/cons to both approaches to the gclient project manangment. I believe the hook would provide benefit in either case, but there is a tradeoff when considering separate gclient projects vs a single one.

A single gclient project provides the simplicity of a single call to `gclient sync` updating all dependencies, but separate gclient projects allows the developer a bit more control over what dependencies get updated: i.e. if I only occasionally full build, it might be annoying to force updates to the entire chromium/src project when I only want to update the devtools dependencies.

I'd like to provide the option for developers to choose their approach, as we also provide approaches for standalone and integrated workflows, but to Yang's point, that might be confusing or harder to maintain.

Perhaps it would be worth an email to devtools-dev to gather feedback? And maybe I could scope this change down to just adding the hook (and readme entry for the hook), which is valuable for either case? The follow up would just be updating the readme/contributing.md with the preferred (or both) gclient option(s).

Patch-set: 7