Also copy over integration_test_runner.html in BUILD.gn
This includes `integration_test_runner.html` in the copy action of
`front_end`. We previously ran into issues with the layout test
runners, which was probably because we erroneously removed the
mention of `integration_test_runner.html` in the `devtools_test_files`.

Therefore, remove all logic related to `has_html` and remove
`integration_test_runner.html` from the root BUILD.gn, but not from
the `devtools_test_files` array which is used on CQ.

R=aerotwist@chromium.org

Fixed: 1078770
Change-Id: Id5379903278fcee569662dc986b64d665c40f3e8