Update patch set 13
Patch Set 13:

(7 comments)

Thanks for trying it out. I added the implementation of closeTabs in the delegate class.

The Console is a special case, since the drawer view is implemented as a simple Wrapper class that shows the same view than the main Console tab. I don't think it is very useful to have 2 Console tabs on the main panel. 
One solution could be to not show the option to move the Console tab. Will defer this UX decision to Peter and Rachel.

I would like to add the new e2e tests in the same file as the cross tool integration tests, waiting for that CL to be approved.

Patch-set: 13