[Dom Storage]: Handle undefined storage case
Issue:
- DomStorageModel.clearForOrigin does not protect against the case where _storages[key] is undefined

Changes:
- Adding !storage check to clearForOrigin

Description:
There are a ton of moving parts in this code.  Navigation, frame addition and frame removal all trigger a series of separate events that change the main security origin, remove unused origins and add new origins.
The failure is occurs when ClearStorageView._securityOrigin has a mismatch of what origins are in the DOMStorageModel._storages map at the time the user clears the site data.

We don't need to worry about the mismatch in the fix since SecurityOriginManager.updateSecurityOrigins will eventually resync the available origins and their storages.
However, we want to return early when we do reach a missing storage in clearForOrigin since the following code will try removing it (resulting in noop) and then add it (will add a storage that wasn't there previously).

Bug: 1084631
Change-Id: I7bb1abe9182e83239d63478a1c6b68feb7d6451e