Update patch set 2
Patch Set 2:

Sorry it took awhile to get back. Just had a chance for a closer look. So, on local testing, it looks like we call calculateAndEmulate and onFrameResized on every switch, which means we are sending 2 setShowHinge calls on every resize.

I guess this is ok, but I'm wondering why we have both calls? are there any cases where we call one and not the other?

Patch-set: 2
Reviewer: Gerrit User 1341669 <1341669@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>