Auto resume all targets reloaded after crash
Since DevTools front-end is auto pausing all targets on start
it is responsible for resuming them. One of the scenarious
when a target has to be resumed is when it's reloaded after
crash. This is the case e.g. for service workers whose life
cycle looks like this:
Target.targetCreated (a new version of sw is created and started)
Inspector.targetCrashed (when the worker is stopped)
Inspector.targetReloadedAfterCrash (when the sw restarts)
Target.targetDestroyed (when the version of sw is stopped and replaced by a new one)

In such cases we don't want to show the crash dialog but are still
oliged to resume the target.

Bug: 1085150
Change-Id: Idb36b50bbeb09632cd8f1480d2f36b24387b7f2e