Remove protobuf dependency in the symbol server
Protobuf is used to generate the JSON-RPI interface from a proto file.
Carrying that dependency around brings extra difficulties on windows and
on the builder infra, so it's worthwile to drop it. Instead of proto,
this CL uses PDL as a spec format and adds simple tools to generate the
bindings.

Bug: chromium:1082696
Change-Id: I81052921f692b4c8dd656da84f5b0f02e9be38b0