[bindings] Make `DebuggerSourceMapping` async.
In order to fix the racy behavior we see in chromium:1056632, we need to
be able to wait on the Wasm bytecode to load and be preprocessed by the
WasmParser before we can translate between UI and raw locations. This CL
makes the relevant functions `async` in preparation for the actual
changes.

Bug: chromium:1032016, chromium:1056632
Change-Id: Ica18f85679e48d856c53036a0efc17c61a579295