Update patch set 7
Patch Set 7:

Had a quick chat with Yang. Our conclusion is that this change simplifies the integration with Chromium and therefore we should adopt it.

Ideally we end up with 2 workflows: 1 for an engineer solely working on DevTools frontend and 1 for an engineer working on the backend, potentially requiring changes to the frontend. For the latter, we currently have 3 workflows (symlink, no symlink, debug_devtools) and this would introduce a fourth one (symlink v2).

We should replace the other workflows, to make sure we converge on one. To do so, we will need to define a migration, write a doc and collect feedback on devtools-dev@ to figure out if we overlooked something.

Patch-set: 7
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>