Rewrite ElementsBreadcrumbs as a TypeScript custom element
This is the first attempt at building a piece of UI in the devtools with
TypeScript and custom elements. This CL also explores how we can interface
between the old code and the new, using events as required.

The _bridge.js file is auto-generated using the component bridges generator.

DISABLE_THIRD_PARTY_CHECK=needed a ts library change

Change-Id: I2ad7d3ab52606b92aa23438f259a253812f55a83