[ElementsTreeElement]: Adding broken state checks before ElementsTreeElement.resize to avoid unexpected failures
Issue:
- The Elements panel can reach a broken editing state which causes failures inside the resize function

Changes:
- Adding a broken editing check before calling resize

Description:
We can reach a broken editing state when the Element is unbound while editing the element as HTML.  In the general case, the element will be unbound and deleted from the DOM tree if the polling finds that the element has been removed. However, if the element is unbound and the ElementsTreeElement.dispose() is called, we can reach a broken state.

Since the element's treeOutline is null when unbound, ElementsTreeOutline._editing is deleted, but the treeOutline's _multilineEditing is not cleared (this is because the element's link to the treeOutline is cut when unbound).
TreeOutline relies on _multilineEditing to see if there is currently an editor open.  In the broken state, it will try to resize the editor window of the unbound element.  This causes failures inside the resize function.



Change-Id: I383f062ebf43527b739f09bd3f44f04fe7cff52a