Synchronize Handling of Paused with Parsed Events for Debugger Plugins
Pausing on breakpoints with language plugins runs into racy situations
when the Paused event arrives before the handling the scriptParsed event
is complete.  Since the scriptParsed handler only hooks the plugins up
once they're fully loaded, pausing misses plugins that are in the
process of loading. Track in progress plugins instead, so that pausing
can wait for them to complete.

Bug: chromium:1084930
Change-Id: I2ffc313fe04b29634f7376de34fad29b261a7d2c