[DevTools] [Emulation] [E2E tests] E2E emulation test helpers
E2E tests for the emulation mode require a different frontend than most
of the other tests. This CL contains helper methods and a simplest test
for device emulation (that the Toggle device mode button is clickable).

Subsequent PRs will add more tests for emulation.

Implementation wise, accessing the correct frontend requires support
from Puppeteer. A hack was suggested and used; we wrote a separate
set of general helper methods ($, waitFor, etc) based on the hack. We
put these methods under emulation because a) they are hackish, and b)
they are only used by emulation. A pointer to the Puppeteer issue where
the hack is discussed is included in the comments.

Change-Id: I957ed0fa107b3b7cac57e222da273371986d21ce