Update patch set 1
Patch Set 1:

> It is likely we will introduce more scripts in the future, which would all then require to have these flags. Solving it in DEPS would allow us to prevent that necessity.

Tying unit test running to DEPS is relevant for the build checkin gate system, but I find this change to be useful for another reason: if I need to run unit tests locally, having this flag allows me to run and compare two different builds of Chromium (for example, public published stable vs. my local build). 

The notion that "more scripts in the future" would be required to have this flag isn't true. Software all starts from zero. Perhaps one day we can add it in to those hypothetical future scripts when the utility for that flag comes around.

> Unit tests are intended to be DevTools only, irrespective of Chromium backend implementation. Therefore, I don't think it is intentional to allow for custom binaries for unit tests. We only use `third_party/chrome` as we already have that binary checked in into the codebase, to make sure test runs on different machines show the same behavior. Other than that, we don't really depend on any Chromium backend characteristics.

It seems to me that this statement makes the argument for this flag stronger, not weaker. Tying unit test running specifically to DEPS introduces a stronger coupling between the unit tests and a particular binary. That combination is fine for the automated tests, but it's unnecessary for manual test runs.

Patch-set: 1
CC: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>