Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Sorry it took awhile to get back. Just had a chance for a closer look. So, on local testing, it looks like we call calculateAndEmulate and onFrameResized on every switch, which means we are sending 2 setShowHinge calls on every resize.
> 
> I guess this is ok, but I'm wondering why we have both calls? are there any cases where we call one and not the other?

I tried with only one show hinge (in onFrameResized, as it is needed for switching device mode on and off). It seems to work like you predicted, except that in debug version at least, the delay between calculateAndEmulate and onFrameResized is long enough to cause problem. If we rotate, for example, a displaced hinge will be visible. So I suggest we keep as is.

Patch-set: 2