Update patch set 1
Patch Set 1: Code-Review+1

Approving to unblock, but agreed with Yang that it feels like this is working around the problem of not having a pinned version of Edge in your fork downstream. Let's get this in for now and revisit later, to clean up the flags if we need to.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1