Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> I'm concerned that we don't adequately understand what's going on with this. There's a 7-line delta from the previous reference to this.treeOutline.something.

I had a more in depth description in my previous patchset where I swapped the ordering of the code.  Here is the description: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2213805/3//COMMIT_MSG

The TLDR is that the this.listItemElement.focus() call is triggering a blur event that will unbind the tree element.  It is the only reasonable explanation as it is the only call between the last valid call to treeoutline and the failing line that would possibly unbind the tree element.

My original fix was to swap the code such that we would dispatch the ElementSelected event before the focus change: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2213805/3/front_end/ui/Treeoutline.js.  After Stan's comment, I figured it would be a safer fix to just add a null check since they also use it in other places in the file.

Patch-set: 6