Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6: Code-Review-1
> > 
> > I'm concerned that we don't adequately understand what's going on with this. There's a 7-line delta from the previous reference to this.treeOutline.something.
> 
> I had a more in depth description in my previous patchset where I swapped the ordering of the code.  Here is the description: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2213805/3//COMMIT_MSG
> 
> The TLDR is that the this.listItemElement.focus() call is triggering a blur event that will unbind the tree element.  It is the only reasonable explanation as it is the only call between the last valid call to treeoutline and the failing line that would possibly unbind the tree element.
> 
> My original fix was to swap the code such that we would dispatch the ElementSelected event before the focus change: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2213805/3/front_end/ui/Treeoutline.js.  After Stan's comment, I figured it would be a safer fix to just add a null check since they also use it in other places in the file.

Your characterization of this problem makes it sound as if some specialization of the TreeOutline is violating an invariant. We shouldn't have to check whether `this.treeOutline` still contains a valid reference after any operation.

The proper fix here seems to be to correct the specialization of the TreeOutline in question, so that it queues an update causing the TreeElement to be unbound. That way the TreeElement's select() operation can be handled atomically.

Patch-set: 6
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>