[TreeElement]: null treeOutline exception in TreeElement.select
Issue:
- this.treeOutline is null at this.treeOutline.dispatchEventToListeners(Events.ElementSelected, this) in TreeElement.select
- This failure occurs in the WatchExpressionsSidebarPane because of the focus shift in TreeElement.select

Repro Steps - GIF: https://imgur.com/vUnbBJQ
- Add watch expression with children
- Edit watch expression and delete all the text
- Click on a child

Clicking on the child will call TreeElement.select which shifts focus to that child's listItemElement.  This triggers the 'blur' event listener on the editor which commits the text change.  If the text is empty, it will delete the TreeElement and unbind all its children.

Changes:
- Collapsing the children when editing a watch expression
- GIF: https://imgur.com/Nh1EoRQ

Bug: 1085816
Change-Id: I1211013d1728154916a62573c84fd3a3af9467f9