DevTools: Use the new grid highlight overlay behind an experiment.
This CL introduces a new runtime flag that defaults to false.
Once turned on, the on-hover highlight will make use of the new grid
overlay for grid elements.
The configuration used to do the highlight is simple and further changes
will happen there, but this is to get us started, allow to iterate and
gather feedback.

Bug: 1047356
Change-Id: I5e0af70b44ae0254a4f4a09d2f1e7746e4df64f7