Update patch set 2
Patch Set 2:

Quick note about this: the configuration defines a hatch color for gaps, but no background color. With the way the overlay code is setup, this means gaps will remain transparent. This is only temporary as I have another CL I'm working on that will make it possible to define a hatch color independently of the background color.

Patch-set: 2