[WIP] Components dev server
This introduces a prototype of a developer server that we would use to view components locally.

This won't be landed in its current form.

The process would be:

1. Create your component.
2. Create `front_end/component_docs/your-component` folder.
3. Add examples as self contained HTML file in `front_end/component_docs/your-component/X.html`
4. Update `front_end/component_docs/BUILD.gn` with references to those HTML files.
5. Run `node scripts/component_server/server.js` (we'd map this to an `npm` command)
6. Visit `localhost:8090/your-component`.
7. See all the examples in one file via iframes.

If we were to accept this proposal I'd like to write the server in TypeScript and tidy up the logging/error handling/etc but hopefully this is enough for the purposes of the proposal.

Change-Id: I6cd7588f045c3cd46e57f1f44441ffa95f0b70bb