[wasm] Display UI locations with hexadecimal pc offsets.
Following the guidelines in [1] we use hexadecimal pc offsets
to display source locations in Wasm modules instead of the line
based locations we use for JavaScript and other text formats.
This pc offset is encoded in the `columnNumber` field (and the
`lineNumber` field is zero) for Wasm locations.

We make the distinction based on the MIME type of the associated
`UISourceCode`, which in case of Wasm is 'application/wasm'.

This partially repairs the Breakpoint view, which is a bit weird
after https://crrev.com/c/2220098 landed before.

[1]: https://github.com/WebAssembly/design/blob/master/Web.md

Bug: chromium:1071432, chromium:1090256
Change-Id: I3e0e774ad8e77a7e1ea6e021887d9fda7785747a