Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Please take a look Brandon. This is the CL that you already looked at before the file move: https://chromium-review.googlesource.com/c/chromium/src/+/2224983
> > 
> > I'm discussed with Peter about this, and I'll post a comment here as soon as we know whether preserving the clipping is required or not.
> 
> I did speak with Peter Muller on Slack, and he confirmed that the approach to not clip anymore made sense.

Sounds good. Thanks for the pictures, that really helps make it clear what this is doing. 

If you could do one more thing: can you post a picture about how this will affect the legacy highlighting? I.e. if we only pass the css_grid color to the highlight. If Peter is ok with the changes to the current version this would cause. I'll sign off, just want to make sure we understand the effects this would have on the current implementation.

If the affects are not acceptible, we can make a backend change to preserve the existing functionality by also passing the css_grid in the grid_info json object, and gating the clipping on that param (or something similar)

Patch-set: 2
Reviewer: Gerrit User 1419292 <1419292@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1341669 <1341669@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>