Fix TempFile on Windows
The LLVM class llvm::sys::fs::TempFile does not work well on Windows.
The problem is that in its implementation of setDeleteDisposition()
TempFile calls SetFileInformationByHandle(...) with
Disposition.DeleteFile = Delete, with the purpose of automatically
deleting the file when it is closed.
However this call prevents the file from being opened again, even by the same
process. It also changes the ownership of the file, and the user that
owns the process does not even own that file anymore.

TempFile is not broken in itself, but it does not work, on Windows, in
the way we want to use it. In particular, we are calling
lld::wasm::link() passing the path of the temp obj file and the path of
the temp wasm module, and link() expects to be able to open these two
temp files to read them.

As a fix, in Variables.cc we could define a simplifies version of class
TempFile that works as we expect on Windows.

Change-Id: I022739e899a32f95fb35a13b262e7c220adb93a3