Update patch set 2
Patch Set 2:

> > If you could do one more thing: can you post a picture about how this will affect the legacy highlighting? I.e. if we only pass the css_grid color to the highlight. If Peter is ok with the changes to the current version this would cause. I'll sign off, just want to make sure we understand the effects this would have on the current implementation.
> > 
> > If the affects are not acceptible, we can make a backend change to preserve the existing functionality by also passing the css_grid in the grid_info json object, and gating the clipping on that param (or something similar)
> 
> granted, we are also intentionally changing the content shading to no longer clip in the next cl, and generally agree the current experiance is not ideal. so I think this effect is also acceptible, just want to make sure Peter has a chance to see this

Here are pictures of how this affects the legacy grid overlay:

Before: https://i.imgur.com/kWDTb1T.jpg
After: https://i.imgur.com/nGTL9hs.jpg

Peter, as Brandon said, please take a look a these and let us know if it is acceptable to have the legacy overlay be changed in this way while we work on the new (v1) overlay.
If not, then we can definitely make protocol/backend changes to avoid it.

My impression is that those changes are minor enough, or actually beneficial, that we don't really need to both with maintaining the legacy overlay as it used to be.

Patch-set: 2