Introducing a general UI interaction test for the Network panel
The Network panel did have 1 e2e test file which covered some very
specific cases such as encoding, mime type and checkbox interaction.
But it did not have a simple and general test for the main UI
interaction flows: displaying the list and selecting requests.

Adding this test means that we now have a place for checks that
verify that the panel works at all.

Bug: 1079076
Change-Id: Idd798d87951393563b0e2c9cd5cf383e5251755f