Update patch set 1
Patch Set 1:

This is take 2 of https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2210313
The first CL landed but later got reverted because the test was flaky.
The reason, I believe, was that waitForSomeRequestsToAppear wasn't actually waiting for the right DOM elements because it wasn't going through shadow DOM with $$.

This CL is almost practically the same as the previous one, only changes are in the helpers file in waitForSomeRequestsToAppear.

Patch-set: 1
Reviewer: Gerrit User 1396442 <1396442@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false