Ungroup same console messages when "Group similar" setting is off
Currently the "Group similar" toggle in the Console setting only applies
to similar but not the same messages. This behavior is confusing to
developers, because this subtle distinction is not shown anywhere.
This CL makes the same messages to be affected by the "Group similar"
setting. Since this setting is by default toggled on, this change would
not spam users by default.

before: https://imgur.com/a/pHv9bc3

after: https://imgur.com/a/Jdp6u9z

Bug: chromium:1082963
Change-Id: Ic377b9b72247c328a6eb0c3da617fa2a9726d9bf