[A11y] fix DOM breakpoint checkbox not announced by screen reader
The DOM breakpoint sidebar items are not identified as checkboxes by ARIAUtils, so screen reader wont announce the checkbox status when the item is focused.
This Cl set the item as checkbox element so the current checkbox status is announced when item is focused, also when pressing space bar to check/uncheck the screen reader would announced the status.

Narrator:
https://imgur.com/a/hSQLdjO

Bug:963183
Change-Id: I1e793bf962fcc21c6c9ebe4e7dd4d44c160a522b