Update patch set 15
Patch Set 15:

Tim, I've seen inconsistent results with timeout of the end to end test on my machine. 
If the DEBUG flag is set, then it runs fine every time. 
If the DEBUG flag is not set, I often get an error that says the 3000ms limit has exceeded, which is weird because I explicitly set it to 10000. This makes me think that it is not overriding the timeout from beforeEach in mocha_hooks.ts correctly.  

Peter and Rachel, per our discussion in the document, I added a special case for both Console tabs. They won't show the context menu to move the tab.

Patch-set: 15