Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> Tim, I've seen inconsistent results with timeout of the end to end test on my machine. 
> If the DEBUG flag is set, then it runs fine every time. 
> If the DEBUG flag is not set, I often get an error that says the 3000ms limit has exceeded, which is weird because I explicitly set it to 10000. This makes me think that it is not overriding the timeout from beforeEach in mocha_hooks.ts correctly.  
> 
> Peter and Rachel, per our discussion in the document, I added a special case for both Console tabs. They won't show the context menu to move the tab.

If the timeout is 3000 seconds, it means that the `beforeEach` has timed out. Not sure how that is possible though, given that the `describe` does not have a `beforeEach`. A timeout in the `describe` is only for test functions in that suite, not for the actual hooks.

Could you link me a CQ run with a timeout so that I can debug?

Patch-set: 15
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>