Update patch set 1
Patch Set 1:

We're in the process of validating in user tests if this feature is really redundant and people aren't finding it at all. When we brought that up in the sync some month ago we wondered if there is telemetry available how many people are using this functionality. There were also concerns that the way it is implemented now isn't accessible as it relies on a hover and not keyboard. The pros of the interface is that it hints on more functionality and could be extended to offer more complex interaction than autocomplete of a single value. For example, this menu could be a place to show a font tool that would change more than one value, or even a grid wizard.

Patch-set: 1
CC: Gerrit User 1391149 <1391149@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>