Adds estimated TBT to timeline
When showing the Total Blocking Time (TBT) value, we rely on the
InteractiveTime event that comes from the trace. However, this event
doesn't always show up for all traces, because Blink may determine
that the InteractiveTime criteria have not been met and therefore not
fire it. In those cases we currently show a message that says
"Unavailable", which is correct but not always that useful to developers.
This CL improves the situation by constructing an estimate of TBT based
on the durations of top level main thread events, which is used when
Blink does not provide a TBT value.

Image: https://imgur.com/a/g0TJu8Q

Change-Id: If304c5e05b885f94cdf1a49dc529c8fc3c14ff56