[wasm] Use base64-arraybuffer decode.
Avoid the `fetch`-trick for the base64 decoding and instead use the
`base64-arraybuffer` library's `decode` function, which is easily 5x
faster for larger Wasm modules. For example in the Tiny3D case from
https://crbug.com/1092466, on a z840 HP workstation running gLinux,
the time to decode the Base64-encoded Wasm bytes drops from 1400ms
to merely 230ms, saving almost 5-10% of the total time.

Bug: 1092466, 1092460, 1058836, 1071432
Change-Id: I75fee9235e6cc69aff3da4e8bd14516e0d3873c6