Update patch set 17
Patch Set 17:

I added a timeout to the beforeEach function since it was needed to enable the experiment after i updated the tests. 
It doesn't seem to be failing in the CQ, it may be something on my machine what causes it to timeout if DEBUG is not set, I'll keep an eye if it happens again.

Patch-set: 17
Label: Commit-Queue=0